Update browser tests to enable features for testing
