+++ test/automated/browser/cypress/support/e2e.js
@@ -15,6 +15,21 @@
 
 // Import commands.js using ES2015 syntax:
 import './commands';
+import fetchData from './fetchData.js';
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
+
+// Put Owncast in a state where it's ready to be tested.
+
+// Set server URL
+fetchData('http://localhost:8080/api/admin/config/serverurl', {
+	method: 'POST',
+	data: { value: 'https://testing.biz' },
+});
+
+// Enable Fediverse features.
+fetchData('http://localhost:8080/api/admin/config/federation/enable', {
+	method: 'POST',
+	data: { value: true },
+});
